+++ tty_darwin.go
@@ -1,0 +1,8 @@
+// +build darwin
+
+package main
+
+const (
+	ioctlReadTermios  = 0x40487413 // syscall.TIOCGETA
+	ioctlWriteTermios = 0x80487414 // syscall.TIOCSETA
+)
